Minor fixes to QCatchSyscalls support
- Refactor syscall number parsing code.
- Add support returning error codes from catch syscall handlers
