+++ src/protocol/common/mod.rs
@@ -1,4 +1,5 @@
 pub mod hex;
+pub mod lists;
 pub mod thread_id;
 
 /// Lightweight wrapper around `&[u8]` which denotes that the contained data is
